transform legacy tests
